requirements.txt
astroid==2.3.2
certifi==2019.9.11
cffi==1.13.1
Click==7.0
colorama==0.4.1
cryptography==2.8
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
future==0.18.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pg8000==1.13.2
psycopg2==2.8.4
pycparser==2.19
pylint==2.4.3
python-telegram-bot==12.2.0
scramp==1.1.0
six==1.12.0
SQLAlchemy==1.3.10
tornado==6.0.3
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
